🚚 Move conceal functions to a new file
